requirements.txt
djangorestframework
django
psycopg2
django-rest-framework